feat(application): improve frontend (style, translations and pagination)
A frontend is served with Bootstrap Flask extension.

Task workers update a read model with only visualization required
information (as the requested text, it's length for ordering,
translation status and, finally, the translated text). Also, a
simple pagination mechanism and a POST route is available for then
implementing an Ajax script to dynamically fetch the translations
and update the table.